requirements.txt
streamlit>=1.31.0
bcrypt>=4.1.2
plotly>=5.18.0
pandas>=2.1.4
reportlab>=4.0.9
Pillow>=10.2.0
typing-extensions>=4.0.0